+++ src/helpers/App.php
@@ -36,6 +36,7 @@ use craft\web\Response as WebResponse;
 use craft\web\Session;
 use craft\web\User as WebUser;
 use craft\web\View;
+use Symfony\Component\Process\PhpExecutableFinder;
 use yii\base\Event;
 use yii\base\InvalidArgumentException;
 use yii\base\InvalidValueException;
@@ -419,6 +420,35 @@ class App
         }
 
         return false;
+    }
+
+    /**
+     * Returns the path to a PHP executable which should be used by sub processes.
+     *
+     * @return string|null The PHP executable path, or `null` if it can’t be determined.
+     * @since 3.9.4
+     */
+    public static function phpExecutable(): ?string
+    {
+        // If PHP_BINARY was set to $_SERVER, update the environment variable to match
+        if (isset($_SERVER['PHP_BINARY']) && $_SERVER['PHP_BINARY'] !== getenv('PHP_BINARY')) {
+            putenv(sprintf('PHP_BINARY=%s', $_SERVER['PHP_BINARY']));
+        }
+
+        if (
+            getenv('PHP_BINARY') === false &&
+            PHP_BINARY &&
+            PHP_SAPI === 'cgi-fcgi' &&
+            str_ends_with(PHP_BINARY, 'php-cgi')
+        ) {
+            // See if a `php` file exists alongside `php-cgi`, and if so, use that
+            $file = dirname(PHP_BINARY) . DIRECTORY_SEPARATOR . 'php';
+            if (@is_executable($file) && !@is_dir($file)) {
+                return $file;
+            }
+        }
+
+        return (new PhpExecutableFinder())->find() ?: null;
     }
 
     /**