+++ src/services/Composer.php
@@ -11,10 +11,10 @@ use Composer\IO\IOInterface;
 use Composer\IO\NullIO;
 use Composer\Json\JsonFile;
 use Craft;
+use craft\helpers\App;
 use craft\helpers\FileHelper;
 use craft\helpers\Json;
 use Symfony\Component\Process\Exception\ProcessFailedException;
-use Symfony\Component\Process\PhpExecutableFinder;
 use Symfony\Component\Process\Process;
 use yii\base\Component;
 use yii\base\Exception;
@@ -203,7 +203,7 @@ class Composer extends Component
         copy(Craft::getAlias('@lib/composer.phar'), $pharPath);
 
         $command = array_merge([
-            (new PhpExecutableFinder())->find() ?: 'php',
+            App::phpExecutable() ?? 'php',
             $pharPath,
         ], $command, [
             '--working-dir',